Rewrite copy, remove unused code, rename a file
